refactor(bulkupload): Refactor buStates
